Check privilege in BlogPostController
- Check privilege in BlogPostController for posting and deleting blog
  posts

+++ src/main/java/fi/tuni/tiko/bloggingsite/blogpost/BlogPostController.java
@@ -1,10 +1,12 @@
 package fi.tuni.tiko.bloggingsite.blogpost;
 
+import fi.tuni.tiko.bloggingsite.LoginController;
 import fi.tuni.tiko.bloggingsite.ResourceCreator;
 import fi.tuni.tiko.bloggingsite.comment.Comment;
 import fi.tuni.tiko.bloggingsite.comment.CommentRepository;
 import fi.tuni.tiko.bloggingsite.exceptions.BlogPostIdNotFoundException;
 import fi.tuni.tiko.bloggingsite.exceptions.CommentNotFoundException;
+import fi.tuni.tiko.bloggingsite.exceptions.UnauthorizedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.Resource;
@@ -15,6 +17,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.HttpClientErrorException;
 
 import static fi.tuni.tiko.bloggingsite.ResourceCreator.createBlogPostResource;
 import static fi.tuni.tiko.bloggingsite.ResourceCreator.createCommentResource;
@@ -33,7 +36,7 @@ public class BlogPostController {
     BlogPostRepository postRepository;
 
     @Autowired
-    CommentRepository commentRepository;
+    LoginController loginController;
 
     @GetMapping("/hello")
     public String hello() {
@@ -41,12 +44,16 @@ public class BlogPostController {
     }
 
     @PostMapping("/post")
-    public Resource<BlogPost> saveBlogPost(@RequestBody BlogPost post) {
-        BlogPost createdPost = postRepository.save(post);
-        Link selfRel = linkTo(methodOn(BlogPostController.class).saveBlogPost(post)).withSelfRel();
-        Link createdRel = linkTo(methodOn(BlogPostController.class).findBlogPostById(createdPost.getId())).withRel("createdBlogPost");
+    public Resource<BlogPost> saveBlogPost(@RequestBody BlogPost post) throws UnauthorizedException {
+        if (loginController.userIsAdmin()) {
+            BlogPost createdPost = postRepository.save(post);
+            Link selfRel = linkTo(methodOn(BlogPostController.class).saveBlogPost(post)).withSelfRel();
+            Link createdRel = linkTo(methodOn(BlogPostController.class).findBlogPostById(createdPost.getId())).withRel("createdBlogPost");
 
-        return new Resource<>(createdPost, selfRel, createdRel);
+            return new Resource<>(createdPost, selfRel, createdRel);
+        } else {
+            throw new UnauthorizedException();
+        }
     }
 
     @GetMapping("/posts")
@@ -81,8 +88,12 @@ public class BlogPostController {
     }
 
     @DeleteMapping("posts/{id}/delete")
-    public void deleteBlogPostById(@PathVariable Long id) throws BlogPostIdNotFoundException {
-        BlogPost post = findBlogPostById(id).getContent();
-        postRepository.delete(post);
+    public void deleteBlogPostById(@PathVariable Long id) throws BlogPostIdNotFoundException, UnauthorizedException {
+        if (loginController.userIsAdmin()) {
+            BlogPost post = findBlogPostById(id).getContent();
+            postRepository.delete(post);
+        } else {
+            throw new UnauthorizedException();
+        }
     }
 }
